[oap-native-sql] move appender class and fix date type
